move result_serializer test to proper location in test_flow_results.py
